requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
aiosqlite==0.19.0
pyzmq==25.1.1
websockets==12.0
